requirements.txt
######conda install######
# conda==4.14.0 
# conda-build==3.22.0
# conda-content-trust==0+unknown
# conda-pack==0.6.0
# conda-package-handling==1.8.1
# conda-repo-cli==1.0.4
# conda-token==0.3.0
# conda-verify==3.4.2
# navigator-updater==0.2.1
# clip==1.0 
# cytoolz==0.11.0 
# executing==0.8.3
# faiss==1.7.1
# pycurl==7.44.1
# TBB==0.2
# torch==1.9.1 torchvision==0.10.1 torchaudio==0.9.1
# zope.event==4.5.0
# zope.interface==5.4.0
######conda install######


aiosignal==1.2.0
alabaster==0.7.12
altair==4.2.0
#anaconda-client==1.9.0
#anaconda-navigator==2.1.1
#anaconda-project==0.10.2
anyio==3.5.0
appdirs==1.4.4
argh==0.26.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.2
asn1crypto==1.4.0
astor==0.8.1
astroid==2.6.6
#astropy==5.0.4
asttokens==2.0.5
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
atomicwrites==1.4.0
attrs==21.4.0
autopep8==1.6.0
axial-positional-embedding==0.2.1
Babel==2.9.1
backcall==0.2.0
backports.functools-lru-cache==1.6.4
backports.shutil-get-terminal-size==1.0.0
backports.tempfile==1.0
backports.weakref==1.0.post1
base58==2.1.1
beautifulsoup4==4.11.1
binaryornot==0.4.4
bitarray==2.4.1
bkcharts==0.2
black==19.10b0
bleach==4.1.0
blinker==1.4
bokeh==2.4.2
boto==2.49.0
boto3==1.20.16
botocore==1.23.16
Bottleneck==1.3.4
brotlipy==0.7.0
cached-property==1.5.2
cachetools==4.2.4
certifi==2022.6.15.1
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.4
cloudpickle==2.0.0
clyent==1.2.2
colorama==0.4.4
contextlib2==0.6.0.post1
cookiecutter==1.7.3
cryptography==3.4.8
cssselect==1.1.0
cvxpy==1.2.0
cycler==0.11.0
Cython==0.29.28
daal4py==2021.5.0
dask==2022.2.1
datasets==1.18.3
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
diff-match-patch==20200713
dill==0.3.4
distributed==2022.2.1
docker-pycreds==0.4.0
docutils==0.17.1
ecos==2.0.10
einops==0.4.0
emoji==1.7.0
entrypoints==0.4
et-xmlfile==1.1.0
fastcache==1.1.0
fastjsonschema==2.15.1
feedfinder2==0.0.4
feedparser==6.0.8
filelock==3.6.0
flake8==3.9.2
Flask==1.1.2
Flask-Cors==3.0.10
flatbuffers==2.0
fonttools==4.25.0
frozenlist==1.3.0
fsspec==2022.2.0
ftfy==6.1.1
future==0.18.2
gast==0.4.0
gbml==1.1.2
gevent==21.8.0
gitdb==4.0.9
GitPython==3.1.26
glob2==0.7
gmpy2==2.1.2
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet==1.1.1
grpcio==1.42.0
h5py==3.6.0
HeapDict==1.0.1
higher==0.2.1
html5lib==1.1
huggingface-hub==0.4.0
hypopt==1.0.9
idna==3.3
imagecodecs==2021.8.26
imageio==2.9.0
imagesize==1.3.0
importlib-metadata==4.11.3
inflection==0.5.1
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==6.9.1
ipython==8.3.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
isort==5.9.3
itsdangerous==2.0.1
jdcal==1.4.1
jedi==0.18.1
jeepney==0.7.1
jieba3k==0.35.1
Jinja2==2.11.3
jinja2-time==0.2.0
jmespath==0.10.0
joblib==1.1.0
json-lines==0.5.0
json5==0.9.6
jsonschema==4.4.0
jupyter==1.0.0
jupyter-client==6.1.12
jupyter-console==6.4.0
jupyter-core==4.10.0
jupyter-server==1.13.5
jupyterlab==3.3.2
jupyterlab-pygments==0.1.2
jupyterlab-server==2.10.3
jupyterlab-widgets==1.0.0
kaggle==1.5.12
keras==2.7.0
Keras-Preprocessing==1.1.2
keyring==23.4.0
kiwisolver==1.3.2
latexcodec==2.0.1
lazy-object-proxy==1.6.0
libarchive-c==2.9
libclang==12.0.0
linear-attention-transformer==0.19.1
linformer==0.2.1
llvmlite==0.38.0
local-attention==1.4.3
locket==0.2.1
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.3.1
mkl-random==1.2.2
mkl-service==2.4.0
mock==4.0.3
monty==2022.4.26
more-itertools==8.12.0
mpmath==1.2.1
msgpack==1.0.2
multidict==6.0.2
multipledispatch==0.6.0
multiprocess==0.70.12.2
munkres==1.1.4
mypy-extensions==0.4.3
nbclassic==0.3.5
nbclient==0.5.13
nbconvert==6.4.4
nbformat==5.3.0
nest-asyncio==1.5.5
networkx==2.7.1
newspaper3k==0.2.8
nltk==3.7
nose==1.3.7
notebook==6.4.11
numba==0.55.1
numexpr==2.8.1
numpy==1.21.5
numpydoc==1.2
oauthlib==3.2.0
olefile==0.46
opencv-python==4.5.5.64
openprompt==1.0.1
openpyxl==3.0.9
opt-einsum==3.3.0
ordered-set==4.1.0
osqp==0.6.2.post5
packaging==21.3
palettable==3.3.0
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
partd==1.2.0
path==16.2.0
pathlib2==2.3.6
pathspec==0.7.0
pathtools==0.1.2
patsy==0.5.2
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
pip==21.2.4
pkginfo==1.8.2
plotly==5.10.0
pluggy==1.0.0
ply==3.11
poyo==0.5.0
product-key-memory==0.1.10
progressbar2==4.0.0
prometheus-client==0.13.1
promise==2.3
prompt-toolkit==3.0.20
protobuf==3.19.1
psutil==5.8.0
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pyarrow==7.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybtex==0.24.0
pycocotools==2.0.4
pycodestyle==2.7.0
pycosat==0.6.3
pycparser==2.21
pydeck==0.7.1
pydocstyle==6.1.1
pyerfa==2.0.0
pyflakes==2.3.1
Pygments==2.11.2
PyJWT==2.1.0
pylint==2.9.6
pyls-spyder==0.4.0
pymatgen==2022.8.23
Pympler==1.0.1
pyodbc==4.0.32
pyOpenSSL==21.0.0
pyparsing==3.0.4
pyrsistent==0.18.0
PySocks==1.7.1
pytest==7.1.1
python-dateutil==2.8.2
python-lsp-black==1.0.0
python-lsp-jsonrpc==1.0.0
python-lsp-server==1.2.4
python-slugify==5.0.2
python-utils==3.3.3
pytorch-nlp==0.5.0
pytorch-pretrained-bert==0.6.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.3.0
pyxdg==0.27
PyYAML==6.0
pyzmq==22.3.0
QDarkStyle==3.0.2
qdldl==0.1.5.post2
qstylizer==0.1.10
QtAwesome==1.0.3
qtconsole==5.3.0
QtPy==2.0.1
regex==2022.3.15
requests==2.27.1
requests-file==1.5.1
requests-oauthlib==1.3.0
rope==0.22.0
rouge==1.0.0
rsa==4.8
Rtree==0.9.7
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
ruamel-yaml-conda==0.15.100
s3transfer==0.5.0
sacremoses==0.0.46
scikit-image==0.19.2
scikit-learn==1.0.2
scikit-learn-intelex
scipy==1.7.3
scs==3.2.0
seaborn==0.11.2
SecretStorage==3.3.1
Send2Trash==1.8.0
sentencepiece==0.1.96
sentry-sdk==1.5.5
seqeval==1.2.2
setuptools==65.3.0
sgmllib3k==1.0.0
shortuuid==1.0.8
simplegeneric==0.8.1
simpletransformers==0.63.4
singledispatch==3.7.0
sip==4.19.13
six==1.16.0
smmap==5.0.0
sniffio==1.2.0
snowballstemmer==2.2.0
sortedcollections==2.1.0
sortedcontainers==2.4.0
soupsieve==2.3.1
spglib==2.0.1
Sphinx==4.4.0
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sphinxcontrib-websupport==1.2.4
spyder==5.1.5
spyder-kernels==2.1.3
SQLAlchemy==1.4.32
stack-data==0.2.0
statsmodels==0.13.2
streamlit==1.5.1
svgling==0.3.1
svgwrite==1.4.2
sympy==1.10.1
tables==3.6.1
tabulate==0.8.10
tblib==1.7.0
tenacity==8.0.1
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.0
tensorboardX==2.4.1
tensorflow==2.7.0
tensorflow-estimator==2.7.0
tensorflow-hub==0.12.0
tensorflow-io-gcs-filesystem==0.22.0
tensorlayer==2.2.5
termcolor==1.1.0
terminado==0.13.1
testpath==0.5.0
text-unidecode==1.3
textdistance==4.2.1
threadpoolctl==2.2.0
three-merge==0.1.1
tifffile==2021.7.2
tinycss==0.4
tinysegmenter==0.3
tldextract==3.2.0
tokenizers==0.12.1
toml==0.10.2
tomli==1.2.2
toolz==0.11.2
torchmeta==1.8.0
torchsummary==1.5.1
tornado==6.1
tqdm==4.64.0
traitlets==5.1.1
transformers==4.19.4
tweepy==4.8.0
tweet-preprocessor==0.6.0
twython==3.9.1
typed-ast==1.4.3
typing_extensions==4.1.1
tzdata==2021.5
tzlocal==4.1
ujson==5.1.0
uncertainties==3.1.7
unicodecsv==0.14.1
Unidecode==1.2.0
urllib3==1.26.9
validators==0.18.2
wandb==0.12.10
watchdog==2.1.6
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==0.58.0
Werkzeug==2.0.3
wget==3.2
wheel==0.37.1
whichcraft==0.6.1
widgetsnbextension==3.5.2
wrapt==1.12.1
wurlitzer==3.0.2
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
xmltodict==0.12.0
xxhash==2.0.2
yacs==0.1.8
yapf==0.31.0
yarl==1.7.2
yaspin==2.1.0
zict==2.0.0
zipp==3.7.0

